Use the m-lab/go/host package to parse -hostname flag (#50)
* Use the m-lab/go/host package to parse -hostname flag

This provides a consistent interface for processing a hostname, and
allows us to handle special cases in the `host` package instead of in
this code. For example, the `host` package knows about managed instance
group random suffixes on instance/node names, and can handle that
gracefully for uuid-annotator.

* Uses builtin String() func instead of fmt

The m-lab/go/host package has a built-in String() function which returns
the hostname, minus any suffix. Use this instead of some long
fmt.Sprintf() call.