Adds public IP of LB to localIPs slice (#55)
* Adds public IP of LB to localIPs slice

Virtual machines will be part of managed instance groups, which sit
behind a pass-through TCP load balancer. In such cases, packets arriving
to the machine from the load balancer will have a destination address of
the load balancer's public IP, which is not a local address on the
machine. Without these changes uuid-annotator will refuse to annoate a
connection, since it doesn't recognize either the Src of Dest IP address
as being its own (annotate.FindDirection() fails). These changes cause
siteannotator.New() to append the public IP address from siteinfo to
localIPs if it is a virtual machine.

* Fixes a typo in a comment