Creates -datadir immediately on startup
Currently, the value of the flag -datadir is only created when
uuid-annotator goes to write the first annotation file. If
uuid-annotator is running in a pod that doesn't get sufficient traffic
which would cause tcp-info to generate an event, then the datatype
directory will not be created. However, pusher will crashloop until this
directory exists. This commit causes uuid-annotator to create the
datatype directory immediately upon startup.