Creates -datadir immediately on startup (#56)
* Creates -datadir immediately on startup

Currently, the value of the flag -datadir is only created when
uuid-annotator goes to write the first annotation file. If
uuid-annotator is running in a pod that doesn't get sufficient traffic
which would cause tcp-info to generate an event, then the datatype
directory will not be created. However, pusher will crashloop until this
directory exists. This commit causes uuid-annotator to create the
datatype directory immediately upon startup.

* Uses mode 755 for MkdirAll() in handler.go

777 is unnecessary, and using 755 is more consistent with the permission
on the base datatype directory, as well as more consistent with how
other sidecar services permission their directories.