Appends loadbalancer v6 addr to localIPs
We were already appending the load balancer v4 address to the localIPs variable
so that uuid-annotator would recognize the non-local, public address of the
machine as "local" and not reject it. This change just adds the IPv6 address,
too, since VMs now support both IPv4 and IPv6.

Additionally, this commit removes the redundant call to parseCIDR(), and just
uses the output from a previous invocation of that same function.