Appends IPv6 address to localIPs in unit test
Additionally, converts the v4 address to a 4 byte representation, which is
apparently the representation of net.IPNet.IP as returned by net.ParseCIDR().
This could be error prone? Maybe we shouldn't be using reflect.DeepEqual() for
comparing slices of IP addresses, since the underlying representation can cause
DeepEqual() to fail, even when the addresses are actually the same.