Append IPv6 address to localIPs (#63)
* Appends loadbalancer v6 addr to localIPs

We were already appending the load balancer v4 address to the localIPs variable
so that uuid-annotator would recognize the non-local, public address of the
machine as "local" and not reject it. This change just adds the IPv6 address,
too, since VMs now support both IPv4 and IPv6.

Additionally, this commit removes the redundant call to parseCIDR(), and just
uses the output from a previous invocation of that same function.

* Appends IPv6 address to localIPs in unit test

Additionally, converts the v4 address to a 4 byte representation, which is
apparently the representation of net.IPNet.IP as returned by net.ParseCIDR().
This could be error prone? Maybe we shouldn't be using reflect.DeepEqual() for
comparing slices of IP addresses, since the underlying representation can cause
DeepEqual() to fail, even when the addresses are actually the same.